move getContributionCartValueData into selector module
