+++ support-frontend/assets/helpers/redux/checkout/product/contributionsSideEffects.ts
@@ -1,11 +1,5 @@
 import { isAnyOf } from '@reduxjs/toolkit';
-import type { ContributionType } from 'helpers/contributions';
-import type { IsoCurrency } from 'helpers/internationalisation/currency';
-import { getContributionType } from 'helpers/redux/checkout/product/selectors/productType';
-import type {
-	ContributionsStartListening,
-	ContributionsState,
-} from 'helpers/redux/contributionsStore';
+import type { ContributionsStartListening } from 'helpers/redux/contributionsStore';
 import * as storage from 'helpers/storage/storage';
 import { trackComponentClick } from 'helpers/tracking/behaviour';
 import { sendEventContributionCartValue } from 'helpers/tracking/quantumMetric';
@@ -16,6 +10,7 @@ import {
 	setProductType,
 	setSelectedAmount,
 } from './actions';
+import { getContributionCartValueData } from './selectors/cartValue';
 
 const shouldCheckFormEnabled = isAnyOf(
 	setAllAmounts,
@@ -28,28 +23,6 @@ const shouldSendEventContributionCartValue = isAnyOf(
 	setProductType,
 	setSelectedAmount,
 );
-
-function getContributionCartValueData(contributionsState: ContributionsState): {
-	contributionAmount: string | number | null;
-	contributionType: ContributionType;
-	contributionCurrency: IsoCurrency;
-} {
-	const pageState = contributionsState.page;
-	const selectedAmounts = pageState.checkoutForm.product.selectedAmounts;
-	const contributionType = getContributionType(contributionsState);
-	const selectedAmount = selectedAmounts[contributionType];
-	const contributionAmount =
-		selectedAmount === 'other'
-			? pageState.checkoutForm.product.otherAmounts[contributionType].amount
-			: selectedAmount;
-	const contributionCurrency = pageState.checkoutForm.product.currency;
-
-	return {
-		contributionAmount,
-		contributionType,
-		contributionCurrency,
-	};
-}
 
 export function addProductSideEffects(
 	startListening: ContributionsStartListening,
